Modification de frontend pour les modifications à faire pour l'admin et le collecteur phase 1

+++ src/main/java/org/example/collectfocep/dto/ClientLocationDTO.java
@@ -1,6 +1,8 @@
 package org.example.collectfocep.dto;
 
 import lombok.*;
+
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Data
@@ -10,8 +12,8 @@ import java.time.LocalDateTime;
 public class ClientLocationDTO {
     private Long clientId;
     private String nomComplet;
-    private Double latitude;
-    private Double longitude;
+    private BigDecimal latitude;
+    private BigDecimal longitude;
     private Boolean coordonneesSaisieManuelle;
     private String adresseComplete;
     private LocalDateTime dateMajCoordonnees;
